refactor(agents): finish exec pipeline split
